refactor: 할 일 체크 이미지 재사용을 위한 분리 및 위치 수정
